[write-fonts] Move IUP code into separate module
I anticipate a bunch more code is going to end up in the variations
module, so I think it's worth trying to split things up into submodules.
